snake & ball game
This program is a game with a snake you control using the arrow keys. The snake must eat the red ball to grow and increase the score. if you hit the wall with the snake the game is over and  the score will be shown like any other game.

NOTE: you must have "pygame" installed for this to work.